Deduplicate Matrix URI generation code
