+++ src/open/clients/Fractal.js
@@ -15,6 +15,7 @@ limitations under the License.
 */
 
 import {Maturity, Platform, LinkKind, FlathubLink} from "../types.js";
+import {Link} from "../../Link.js"
 
 /**
  * Information on how to deep link to a given matrix client.
@@ -31,29 +32,7 @@ export class Fractal {
 
     getDeepLink(platform, link) {
         if (platform === Platform.Linux) {
-            let identifier = encodeURIComponent(link.identifier.substring(1));
-            let isRoomid = link.identifier.substring(0, 1) === '!';
-            let fragmentPath;
-            switch (link.kind) {
-                case LinkKind.User:
-                    fragmentPath = `u/${identifier}?action=chat`;
-                    break;
-                case LinkKind.Room:
-                case LinkKind.Event:
-                    if (isRoomid)
-                        fragmentPath = `roomid/${identifier}`;
-                    else
-                        fragmentPath = `r/${identifier}`;
-
-                    if (link.kind === LinkKind.Event)
-                        fragmentPath += `/e/${encodeURIComponent(link.eventId.substring(1))}`;
-                    fragmentPath += '?action=join';
-                    fragmentPath += link.servers.map(server => `&via=${encodeURIComponent(server)}`).join('');
-                    break;
-                case LinkKind.Group:
-                    return;
-            }
-            return `matrix:${fragmentPath}`;
+            return Link.toMatrixUri(link);
         }
     }
